move app insights to appInsights file
