+++ src/index.tsx
@@ -1,46 +1,16 @@
 import './index.module.scss';
 import 'antd/dist/antd.css';
 
-import {
-  AppInsightsContext,
-  ReactPlugin,
-  withAITracking,
-} from '@microsoft/applicationinsights-react-js';
-import {
-  ApplicationInsights,
-  IExceptionTelemetry,
-} from '@microsoft/applicationinsights-web';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { createBrowserHistory } from 'history';
-import React from 'react';
-import ReactDOM from 'react-dom';
-
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { AppInsightsContext, ReactPlugin } from '@microsoft/applicationinsights-react-js';
+import { IExceptionTelemetry } from '@microsoft/applicationinsights-web';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { getAppInsights } from './appInsights';
 
-// #region Application insight
-const browserHistory = createBrowserHistory({});
 const reactPlugin = new ReactPlugin();
-export default withAITracking(reactPlugin, App);
-export const appInsights = new ApplicationInsights({
-  config: {
-    instrumentationKey: process.env.REACT_APP_APPINSIGHTS_KEY,
-    enableAutoRouteTracking: true,
-    extensions: [reactPlugin],
-    extensionConfig: {
-      [reactPlugin.identifier]: { history: browserHistory },
-    },
-  },
-});
-
-appInsights.loadAppInsights();
-
-window.addEventListener('unhandledrejection', (event) => {
-  // eslint-disable-next-line no-console
-  console.log(event);
-  appInsights.trackException({ error: event.reason } as IExceptionTelemetry);
-});
-// #endregion
+let appInsights = getAppInsights( reactPlugin );
 
 try {
   ReactDOM.render(
@@ -51,8 +21,8 @@ try {
     </React.StrictMode>,
     document.getElementById('root')
   );
-} catch (err: unknown) {
-  appInsights.trackException({ error: err } as IExceptionTelemetry);
+} catch (error: unknown) {
+  appInsights.trackException({ error } as IExceptionTelemetry);
 }
 
 // If you want to start measuring performance in your app, pass a function
